Move project files to Content folders Add VS templating steps
