+++ templates/FluentUIBlazorServerApp/Content/FluentUIBlazorServerApp/.template.config/template.json
