+++ templates/FluentUIBlazorServerApp/Content/FluentUIBlazorServerApp/App.razor
