+++ templates/FluentUIBlazorServerApp/Content/FluentUIBlazorServerApp/Data/WeatherForecastService.cs
@@ -1,0 +1,19 @@
+namespace FluentUIBlazorServerApp.Data;
+
+public class WeatherForecastService
+{
+    private static readonly string[] Summaries = new[]
+    {
+        "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
+    };
+
+    public Task<WeatherForecast[]> GetForecastAsync(DateTime startDate)
+    {
+        return Task.FromResult(Enumerable.Range(1, 5).Select(index => new WeatherForecast
+        {
+            Date = startDate.AddDays(index),
+            TemperatureC = Random.Shared.Next(-20, 55),
+            Summary = Summaries[Random.Shared.Next(Summaries.Length)]
+        }).ToArray());
+    }
+}
