+++ templates/FluentUIBlazorServerApp/Content/FluentUIBlazorServerApp/FluentUIBlazorServerApp.csproj
@@ -1,0 +1,14 @@
+<Project Sdk="Microsoft.NET.Sdk.Web">
+
+  <PropertyGroup>
+	<TargetFramework>net6.0</TargetFramework>
+	<Nullable>enable</Nullable>
+	<ImplicitUsings>enable</ImplicitUsings>
+	<IsPackable>true</IsPackable>
+  </PropertyGroup>
+
+  <ItemGroup>
+	<PackageReference Include="Microsoft.Fast.Components.FluentUI" Version="1.1.0" />
+  </ItemGroup>
+
+</Project>
