+++ templates/FluentUIBlazorServerApp/Content/FluentUIBlazorServerApp/Pages/Counter.razor
