+++ templates/FluentUIBlazorServerApp/Content/FluentUIBlazorServerApp/Program.cs
