+++ templates/FluentUIBlazorServerApp/Content/FluentUIBlazorServerApp/Properties/launchSettings.json
