+++ templates/FluentUIBlazorServerApp/Content/FluentUIBlazorServerApp/Shared/NavMenu.razor
