+++ templates/FluentUIBlazorServerApp/Content/FluentUIBlazorServerApp/appsettings.json
