+++ templates/FluentUIBlazorServerApp/FluentUIBlazorServerApp.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.1.32102.18
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FluentUIBlazorServerApp", "FluentUIBlazorServerApp\FluentUIBlazorServerApp.csproj", "{9A828C59-809E-49E0-A4B3-034E9FAB82BE}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FluentUIBlazorServerApp", "Content\FluentUIBlazorServerApp\FluentUIBlazorServerApp.csproj", "{9A828C59-809E-49E0-A4B3-034E9FAB82BE}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
