--- templates/FluentUIBlazorServerApp/FluentUIBlazorServerApp/FluentUIBlazorServerApp.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Fast.Components.FluentUI" Version="1.1.0" />
-  </ItemGroup>
-
-</Project>
