+++ templates/FluentUIBlazorWebAssemblyApp/Content/FluentUIBlazorWebAssemblyApp/.template.config/ide.host.json
@@ -1,0 +1,5 @@
+{
+  "$schema": "http://json.schemastore.org/vs-2017.3.host",
+  "order": 600,
+  "icon": "icon.png"
+}
