+++ templates/FluentUIBlazorWebAssemblyApp/Content/FluentUIBlazorWebAssemblyApp/App.razor
