+++ templates/FluentUIBlazorWebAssemblyApp/Content/FluentUIBlazorWebAssemblyApp/FluentUIBlazorWebAssemblyApp.csproj
@@ -1,0 +1,16 @@
+<Project Sdk="Microsoft.NET.Sdk.BlazorWebAssembly">
+
+  <PropertyGroup>
+	<TargetFramework>net6.0</TargetFramework>
+	<Nullable>enable</Nullable>
+	<ImplicitUsings>enable</ImplicitUsings>
+	<IsPackable>true</IsPackable>
+  </PropertyGroup>
+
+  <ItemGroup>
+	<PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly" Version="6.0.1" />
+	<PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.DevServer" Version="6.0.1" PrivateAssets="all" />
+	<PackageReference Include="Microsoft.Fast.Components.FluentUI" Version="1.1.0" />
+  </ItemGroup>
+
+</Project>
