+++ templates/FluentUIBlazorWebAssemblyApp/Content/FluentUIBlazorWebAssemblyApp/Program.cs
