+++ templates/FluentUIBlazorWebAssemblyApp/Content/FluentUIBlazorWebAssemblyApp/wwwroot/sample-data/weather.json
