+++ templates/FluentUIBlazorWebAssemblyApp/Microsoft.Fast.Components.FluentUI.BlazorWebAssembly.Template.nuspec
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<package xmlns="http://schemas.microsoft.com/packaging/2012/06/nuspec.xsd">
+  <metadata>
+    <id>Microsoft.Fast.Components.FluentUI.BlazorWebAssembly.Template</id>
+    <version>1.0.0</version>
+    <description>
+      A project template for creating a Blazor app that uses the Fluent UI component library and runs on WebAssembly. This template can be used for web apps with rich dynamic user interfaces (UIs).
+    </description>
+    <authors>Microsoft Fast</authors>
+    <license type="expression">MIT</license>
+    <packageTypes>
+      <packageType name="Template" />
+    </packageTypes>
+  </metadata>
+  <files>
+    <file src="Content\**\*.*" exclude="Content\**\bin\**\*.*;Content\**\obj\**\*.*" target="Content" />
+  </files>  
+</package>
